add package name checking and replacement
any invalid chars in package name are replaced with a valid character, by default an underscore. by default the number of consecutive replaced chars is not reduced. note that there may be inconsistent behavior with package names with invalid chars, and you are still suggested to rename the package to remove such chars.